- Modified file structure to mirrors our proposed architecture
