+++ RestWCFServiceLibrary/App.config
@@ -11,9 +11,9 @@
   <system.serviceModel>
     <services>
             <service behaviorConfiguration="RestWCFServiceLibrary.ServiceBehavior"
-        name="RestWCFServiceLibrary.HighScoreController">
+        name="RestWCFServiceLibrary.Use_Cases.HighScores.HighScoreController">
         <endpoint address="" behaviorConfiguration="web" binding="webHttpBinding"
-          contract="RestWCFServiceLibrary.IHighScoreController" />
+          contract="RestWCFServiceLibrary.Use_Cases.HighScores.Interfaces.IHighScoreController" />
         <endpoint address="mex" binding="mexHttpBinding" contract="IMetadataExchange" />
         <host>
           <baseAddresses>
@@ -22,9 +22,9 @@
         </host>
       </service>
       <service behaviorConfiguration="RestWCFServiceLibrary.ServiceBehavior"
-        name="RestWCFServiceLibrary.TeacherService">
+        name="RestWCFServiceLibrary.Use_Cases.Teachers.TeachersController">
         <endpoint address="" behaviorConfiguration="web" binding="webHttpBinding"
-          contract="RestWCFServiceLibrary.ITeacherService" />
+          contract="RestWCFServiceLibrary.Use_Cases.Teachers.Interfaces.ITeachersController" />
         <endpoint address="mex" binding="mexHttpBinding" contract="IMetadataExchange" />
         <host>
           <baseAddresses>
