+++ RestWCFServiceLibrary/Application.cs
@@ -1,11 +1,7 @@
 ﻿using Autofac;
 using RestWCFServiceLibrary.Repos;
-using System;
-using System.Collections.Generic;
-using System.Data.SQLite;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using RestWCFServiceLibrary.Use_Cases.HighScores;
+using RestWCFServiceLibrary.Use_Cases.HighScores.Interfaces;
 
 namespace RestWCFServiceLibrary
 {
