--- RestWCFServiceLibrary/ITeacherService.cs
@@ -1,23 +1,0 @@
-﻿using System.ServiceModel;
-using System.ServiceModel.Web;
-
-namespace RestWCFServiceLibrary
-{
-    [ServiceContract]
-    public interface ITeacherService
-    {
-        [OperationContract]
-        [WebInvoke(Method = "GET",
-            ResponseFormat = WebMessageFormat.Xml,
-            BodyStyle = WebMessageBodyStyle.Wrapped,
-            UriTemplate = "xml/{id}")]
-        string XMLData(string id);
-
-        [OperationContract]
-        [WebInvoke(Method = "GET",
-            ResponseFormat = WebMessageFormat.Json,
-            BodyStyle = WebMessageBodyStyle.Wrapped,
-            UriTemplate = "json/{id}")]
-        string JSONData(string id);
-    }
-}
