+++ RestWCFServiceLibrary/RestWCFServiceLibrary.csproj
@@ -75,19 +75,19 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="IHighScoreController.cs" />
-    <Compile Include="ITeacherService.cs" />
-    <Compile Include="Models\HighScore.cs" />
+    <Compile Include="Use Cases\HighScores\Interfaces\IHighScoreController.cs" />
+    <Compile Include="Use Cases\Teachers\Interfaces\ITeachersController.cs" />
+    <Compile Include="Use Cases\HighScores\Models\HighScore.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Repos\DatabaseConnection.cs" />
-    <Compile Include="Repos\HighScoreRepo.cs" />
-    <Compile Include="Repos\IHighScoreRepo.cs" />
-    <Compile Include="Repos\IDatabaseConnection.cs" />
-    <Compile Include="HighScoreController.cs" />
+    <Compile Include="Shared\DatabaseConnection.cs" />
+    <Compile Include="Use Cases\HighScores\HighScoreRepo.cs" />
+    <Compile Include="Use Cases\HighScores\Interfaces\IHighScoreRepo.cs" />
+    <Compile Include="Shared\IDatabaseConnection.cs" />
+    <Compile Include="Use Cases\HighScores\HighScoreController.cs" />
     <Compile Include="Application.cs" />
-    <Compile Include="Services\HighScoreService.cs" />
-    <Compile Include="Services\IHighScoreService.cs" />
-    <Compile Include="TeacherService.cs" />
+    <Compile Include="Use Cases\HighScores\HighScoreService.cs" />
+    <Compile Include="Use Cases\HighScores\Interfaces\IHighScoreService.cs" />
+    <Compile Include="Use Cases\Teachers\TeachersController.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config">
@@ -95,9 +95,7 @@
     </None>
     <None Include="packages.config" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Controllers\" />
-  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
     <VSToolsPath Condition="'$(VSToolsPath)' == '' and '$(VisualStudioVersion)' != '' and '$(VisualStudioVersion)' != '11.0'">$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)</VSToolsPath>
