+++ RestWCFServiceLibrary/Shared/DatabaseConnection.cs
