+++ RestWCFServiceLibrary/Shared/IDatabaseConnection.cs
