--- RestWCFServiceLibrary/TeacherService.cs
@@ -1,20 +1,0 @@
-﻿namespace RestWCFServiceLibrary
-{
-    public class TeacherService : ITeacherService
-    {
-        public string XMLData(string id)
-        {
-            return Data(id);
-        }
-        public string JSONData(string id)
-        {
-            return Data(id);
-        }
-
-        private string Data(string id)
-        {
-            // logic
-            return "Data: " + id;
-        }
-    }
-}
