+++ RestWCFServiceLibrary/Use Cases/HighScores/HighScoreController.cs
@@ -1,11 +1,9 @@
-﻿using RestWCFServiceLibrary.Models;
-using RestWCFServiceLibrary.Repos;
-using RestWCFServiceLibrary.Services;
-using System;
+﻿using RestWCFServiceLibrary.Repos;
+using RestWCFServiceLibrary.Use_Cases.HighScores.Interfaces;
+using RestWCFServiceLibrary.Use_Cases.HighScores.Models;
 using System.Collections.Generic;
-using System.Data.SQLite;
 
-namespace RestWCFServiceLibrary
+namespace RestWCFServiceLibrary.Use_Cases.HighScores
 {
     public class HighScoreController : IHighScoreController
     {
