+++ RestWCFServiceLibrary/Use Cases/HighScores/Interfaces/IHighScoreRepo.cs
@@ -1,0 +1,16 @@
+﻿using RestWCFServiceLibrary.Use_Cases.HighScores.Models;
+using System.Collections.Generic;
+
+namespace RestWCFServiceLibrary.Use_Cases.HighScores.Interfaces
+{
+    internal interface IHighScoreRepo
+    {
+        void Create(HighScore highscore);
+
+        HighScore Read(int id);
+
+        IEnumerable<HighScore> ReadAll();
+
+        void Delete(int id);
+    }
+}
