+++ RestWCFServiceLibrary/Use Cases/HighScores/Models/HighScore.cs
@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Runtime.Serialization;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Runtime.Serialization;
 
-namespace RestWCFServiceLibrary.Models
+namespace RestWCFServiceLibrary.Use_Cases.HighScores.Models
 {
     [DataContract]
     public class HighScore
