+++ WindowsService1/App.config
@@ -10,9 +10,9 @@
   <system.serviceModel>
     <services>
       <service behaviorConfiguration="RestWCFServiceLibrary.ServiceBehavior"
-  name="RestWCFServiceLibrary.HighScoreController">
+        name="RestWCFServiceLibrary.Use_Cases.HighScores.HighScoreController">
         <endpoint address="" behaviorConfiguration="web" binding="webHttpBinding"
-          contract="RestWCFServiceLibrary.IHighScoreController" />
+          contract="RestWCFServiceLibrary.Use_Cases.HighScores.Interfaces.IHighScoreController" />
         <endpoint address="mex" binding="mexHttpBinding" contract="IMetadataExchange" />
         <host>
           <baseAddresses>
@@ -21,9 +21,9 @@
         </host>
       </service>
       <service behaviorConfiguration="RestWCFServiceLibrary.ServiceBehavior"
-        name="RestWCFServiceLibrary.TeacherService">
+        name="RestWCFServiceLibrary.Use_Cases.Teachers.TeachersController">
         <endpoint address="" behaviorConfiguration="web" binding="webHttpBinding"
-          contract="RestWCFServiceLibrary.ITeacherService" />
+          contract="RestWCFServiceLibrary.Use_Cases.Teachers.Interfaces.ITeachersController" />
         <endpoint address="mex" binding="mexHttpBinding" contract="IMetadataExchange" />
         <host>
           <baseAddresses>
